added buffer max retained capacity
