+++ tests/ReadableByteBufferTests.cs
@@ -219,6 +219,126 @@ public class WrappedByteBufferTests
         Assert.Equal( $"{{\"Id\":\"{obj.Id}\",\"Number\":{obj.Number},\"Buffer\":\"000102030405\"}}", json );
     }
 
+    [Fact]
+    public void Test_FindBytes()
+    {
+        var source = new ReadableByteBuffer(
+            [0x00, 0x01, 0x02, 0x03]
+        );
+
+        Assert.Equal( 0, source.IndexOf( [0x00, 0x01] ) );
+        Assert.Equal( 1, source.IndexOf( [0x01, 0x02] ) );
+        Assert.Equal( 2, source.IndexOf( [0x02, 0x03] ) );
+        Assert.Equal( 3, source.IndexOf( [0x03] ) );
+
+        Assert.Equal( 0, source.IndexOf( 0x00 ) );
+        Assert.Equal( 1, source.IndexOf( 0x01 ) );
+        Assert.Equal( 2, source.IndexOf( 0x02 ) );
+        Assert.Equal( 3, source.IndexOf( 0x03 ) );
+
+        source.SkipBytes( 1 );
+
+        Assert.Equal( 1, source.IndexOf( [0x01, 0x02] ) );
+        Assert.Equal( 2, source.IndexOf( [0x02, 0x03] ) );
+        Assert.Equal( 3, source.IndexOf( [0x03] ) );
+        Assert.Equal( 3, source.IndexOf( [0x03], 3 ) );
+
+        Assert.Equal( -1, source.IndexOf( 0x00 ) );
+        Assert.Equal( 1, source.IndexOf( 0x01 ) );
+        Assert.Equal( 2, source.IndexOf( 0x02 ) );
+        Assert.Equal( 3, source.IndexOf( 0x03 ) );
+    }
+
+    [Fact]
+    public void Test_IndexOf()
+    {
+        var buffer = new ReadableByteBuffer( [0x00, 0x01, 0x02, 0x03, 0x04, 0x05] );
+
+        Assert.Equal( -1, buffer.IndexOf( 0xff ) );
+        Assert.Equal( -1, buffer.IndexOf( 0xff, 0 ) );
+        Assert.Equal( -1, buffer.IndexOf( 0xff, 5 ) );
+
+        Assert.Equal( -1, buffer.IndexOf( [0xff] ) );
+        Assert.Equal( -1, buffer.IndexOf( [0xff], 5 ) );
+        Assert.Equal( -1, buffer.IndexOf( [0xff], 0 ) );
+
+        Assert.Equal( -1, buffer.IndexOf( [0xff, 0xff] ) );
+        Assert.Equal( -1, buffer.IndexOf( [0xff, 0xff], 4 ) );
+        Assert.Equal( -1, buffer.IndexOf( [0xff, 0xff], 5 ) );
+        Assert.Equal( -1, buffer.IndexOf( [0xff, 0xff], 0 ) );
+
+        Assert.Equal( 1, buffer.IndexOf( [0x01, 0x02] ) );
+        Assert.Equal( 1, buffer.IndexOf( [0x01, 0x02], 0 ) );
+        Assert.Equal( 1, buffer.IndexOf( [0x01, 0x02], 1 ) );
+
+        Assert.Equal( 0, buffer.IndexOf( [0x00] ) );
+        Assert.Equal( 0, buffer.IndexOf( [0x00], 0 ) );
+        
+        Assert.Equal( 1, buffer.IndexOf( [0x01] ) );
+        Assert.Equal( 1, buffer.IndexOf( [0x01], 0 ) );
+        Assert.Equal( 1, buffer.IndexOf( [0x01], 1 ) );
+
+        Assert.Equal( 2, buffer.IndexOf( [0x02] ) );
+        Assert.Equal( 2, buffer.IndexOf( [0x02], 0 ) );
+        Assert.Equal( 2, buffer.IndexOf( [0x02], 1 ) );
+        Assert.Equal( 2, buffer.IndexOf( [0x02], 2 ) );
+
+        Assert.Equal( 3, buffer.IndexOf( [0x03] ) );
+        Assert.Equal( 3, buffer.IndexOf( [0x03], 0 ) );
+        Assert.Equal( 3, buffer.IndexOf( [0x03], 1 ) );
+        Assert.Equal( 3, buffer.IndexOf( [0x03], 2 ) );
+        Assert.Equal( 3, buffer.IndexOf( [0x03], 3 ) );
+
+        Assert.Equal( 4, buffer.IndexOf( [0x04] ) );
+        Assert.Equal( 4, buffer.IndexOf( [0x04], 0 ) );
+        Assert.Equal( 4, buffer.IndexOf( [0x04], 1 ) );
+        Assert.Equal( 4, buffer.IndexOf( [0x04], 2 ) );
+        Assert.Equal( 4, buffer.IndexOf( [0x04], 3 ) );
+        Assert.Equal( 4, buffer.IndexOf( [0x04], 4 ) );
+
+        Assert.Equal( 5, buffer.IndexOf( [0x05] ) );
+        Assert.Equal( 5, buffer.IndexOf( [0x05], 0 ) );
+        Assert.Equal( 5, buffer.IndexOf( [0x05], 1 ) );
+        Assert.Equal( 5, buffer.IndexOf( [0x05], 2 ) );
+        Assert.Equal( 5, buffer.IndexOf( [0x05], 3 ) );
+        Assert.Equal( 5, buffer.IndexOf( [0x05], 4 ) );
+        Assert.Equal( 5, buffer.IndexOf( [0x05], 5 ) );
+    }
+
+    [Fact]
+    public void Test_Any()
+    {
+        var buffer = new ReadableByteBuffer( [0x00, 0x01, 0x02, 0x03, 0x04, 0x05] );
+
+        Assert.True( buffer.Any( b => b == 0x03 ) );
+        Assert.True( buffer.Any( b => b > 0x00 ) );
+        Assert.False( buffer.Any( b => b == 0xff ) );
+    }
+
+    [Fact]
+    public void Test_UndoRead()
+    {
+        var buffer = new ReadableByteBuffer( [0x00, 0x01, 0x02, 0x03] );
+
+        Assert.Equal( 0x00, buffer.ReadByte() );
+        Assert.Equal( 0x01, buffer.ReadByte() );
+
+        buffer.UndoRead( 1 );
+
+        Assert.Equal( 0x01, buffer.ReadByte() );
+
+        buffer.UndoRead( 2 );
+
+        Assert.Equal( 0x00, buffer.ReadByte() );
+        Assert.Equal( 0x01, buffer.ReadByte() );
+        Assert.Equal( 0x02, buffer.ReadByte() );
+        Assert.Equal( 0x03, buffer.ReadByte() );
+
+        buffer.ResetOffset();
+
+        Assert.Throws<ArgumentOutOfRangeException>( () => buffer.UndoRead( 1 ) );
+    }
+
     private class ObjWithBuffer
     {
         public required string Id { get; set; }
