+++ tests/WritableByteBufferTests.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Buffers.Binary;
 using Faactory.Channels.Buffers;
 using Xunit;
 
@@ -264,5 +265,107 @@ public class WritableByteBufferTests
         buffer.WriteBytes([1, 2, 3]);
 
         Assert.Throws<ArgumentOutOfRangeException>(() => buffer.Truncate(4));
+    }
+
+    [Fact]
+    public void Writable_GrowsCorrectly()
+    {
+        var buffer = new WritableByteBuffer();
+
+        for ( int i = 0; i < 10000; i++ )
+        {
+            buffer.WriteInt32( i );
+        }
+
+        Assert.Equal( 10000 * 4, buffer.Length );
+
+        // get private buffer field using reflection
+        var field = typeof( WritableByteBuffer ).GetField( "buffer", System.Reflection.BindingFlags.NonPublic | System.Reflection.BindingFlags.Instance );
+        var internalBuffer = (byte[])field!.GetValue( buffer )!;
+
+        /*
+        initial capacity is 1024
+        every time the buffer needs to grow, it doubles its capacity
+        writing 4 bytes at a time, the buffer will grow as follows:
+        1024 -> 2048 -> 4096 -> 8192 -> 16384 -> 32768 -> 65536 -> 131072 -> etc...
+        the first 1024 need 256 writes, the next 1024 need 256 writes, the next 2048 need 512 writes, etc...
+        how many times the buffer needs to grow to accommodate 10000 * 4 bytes?
+        10000 * 4 = 40000 bytes
+        1024 -> 2048 -> 4096 -> 8192 -> 16384 -> 32768 -> 65536
+        the buffer needs to grow 6 times to accommodate 40000 bytes, which means the final capacity will be 65536 bytes
+        */
+
+        Assert.Equal( 65536, internalBuffer.Length );
+
+        var span = buffer.AsSpan();
+
+        for ( int i = 0; i < 10000; i++ )
+        {
+            var value = BinaryPrimitives.ReadInt32BigEndian(
+                span.Slice( i * 4, 4 )
+            );
+
+            Assert.Equal( i, value );
+        }
+    }
+
+    [Fact]
+    public void Writable_Compact_ShouldShrinkBuffer_WhenCapacityExceedsMaxRetainedCapacity()
+    {
+        byte[]? releasedBuffer = null;
+
+        var buffer = new WritableByteBuffer(
+            new WritableByteBufferOptions
+            {
+                InitialCapacity = 16,
+                MaxRetainedCapacity = 16
+            },
+            releaser: b => releasedBuffer = b
+        );
+
+        // force growth
+        buffer.WriteBytes( new byte[128] );
+
+        var grownCapacity = buffer.Buffer.Length;
+
+        Assert.True( grownCapacity > 16 );
+
+        // compact most data away
+        buffer.Compact( 120 );
+
+        // remaining data should be small
+        Assert.Equal( 8, buffer.Length );
+
+        // old oversized buffer should have been released
+        Assert.NotNull( releasedBuffer );
+
+        // buffer should shrink back to retained capacity
+        Assert.Equal( 16, buffer.Buffer.Length );
+    }
+
+    [Fact]
+    public void Writable_Compact_WithZeroOffset_ShouldShrinkBuffer_WhenCapacityExceedsMaxRetainedCapacity()
+    {
+        var buffer = new WritableByteBuffer(
+            new WritableByteBufferOptions
+            {
+                InitialCapacity = 16,
+                MaxRetainedCapacity = 32
+            }
+        );
+
+        // force growth
+        buffer.WriteBytes( new byte[128] );
+
+        Assert.True( buffer.Buffer.Length > 32 );
+
+        // simulate logical reset
+        buffer.Truncate();
+
+        // should still apply retention policy
+        buffer.Compact( 0 );
+
+        Assert.Equal( 0, buffer.Length );
+        Assert.Equal( 32, buffer.Buffer.Length );
     }
 }
